Commit 4 - File Reading on Exercise Activity
Added file read and write functionality. Exercise activity will now read and populate the RecyclerView on creation and save button will save to CSV file.

Added function to insert and remove items from the RecyclerView. Items in RecyclerView will now be added or removed to equal the "sets" input field when it is unfocused.