Commit 5 - Main Activity
Application now reading file each time the main activity is resumed/started.

RecyclerView working on Main Activity, showing list of exercises and list of sets for each exercise.

Left and right navigation keys working on the Exercise activity

Added error detection and feedback for empty exercise name

Added customised Application class in order to store the file (when read) in an object that can be accessed throughout the application. Object is updated when the file is saved to; this is so the file doesn't have to be read again at the start of each activity.

Incorporated customised RecyclerView Adapters back into the activity classes specific for each activity and added directories for models and activities; for source file neatness and organisation.

Updated ToDo List